Financial News Sentiment Analysis/requirements.txt
numpy==1.23.5
streamlit==1.15.1
tensorflow_cpu==2.8.0
tensorflow_hub==0.12.0

